+++ SolarCore/SolarObjects/abstractsolarobject.h
@@ -17,63 +17,63 @@ namespace SolarSystem
         virtual ~AbstractSolarObject();
 
         //getters + setters
-        double radius() const;
-        void setRadius(double radius);
+        float radius() const;
+        void setRadius(float radius);
 
-        double tilt() const;
-        void setTilt(double tilt);
+        float tilt() const;
+        void setTilt(float tilt);
 
-        double roll() const;
-        void setRoll(double roll);
+        float roll() const;
+        void setRoll(float roll);
 
-        double x() const;
-        void setX(double x);
+        float x() const;
+        void setX(float x);
 
-        double y() const;
-        void setY(double y);
+        float y() const;
+        void setY(float y);
 
-        double z() const;
-        void setZ(double z);
+        float z() const;
+        void setZ(float z);
 
         //only getters
-        double N1() const;
-        double N2() const;
-        double i1() const;
-        double i2() const;
-        double w1() const;
-        double w2() const;
-        double a1() const;
-        double a2() const;
-        double e1() const;
-        double e2() const;
-        double M1() const;
-        double M2() const;
-        double period() const;
+        float N1() const;
+        float N2() const;
+        float i1() const;
+        float i2() const;
+        float w1() const;
+        float w2() const;
+        float a1() const;
+        float a2() const;
+        float e1() const;
+        float e2() const;
+        float M1() const;
+        float M2() const;
+        float period() const;
         SolarObjects centerOfOrbit() const;
         SolarObjects object() const;
 
     protected:
 
         //main data
-        double m_Radius = 0;
-        double m_Tilt = 0;
-        double m_N1 = 0;
-        double m_N2 = 0;
-        double m_I1 = 0;
-        double m_I2 = 0;
-        double m_W1 = 0;
-        double m_W2 = 0;
-        double m_A1 = 0;
-        double m_A2 = 0;
-        double m_E1 = 0;
-        double m_E2 = 0;
-        double m_M1 = 0;
-        double m_M2 = 0;
-        double m_Period = 0;
-        double m_X = 0;
-        double m_Y = 0;
-        double m_Z = 0;
-        double m_Roll = 0;
+        float m_Radius = 0;
+        float m_Tilt = 0;
+        float m_N1 = 0;
+        float m_N2 = 0;
+        float m_I1 = 0;
+        float m_I2 = 0;
+        float m_W1 = 0;
+        float m_W2 = 0;
+        float m_A1 = 0;
+        float m_A2 = 0;
+        float m_E1 = 0;
+        float m_E2 = 0;
+        float m_M1 = 0;
+        float m_M2 = 0;
+        float m_Period = 0;
+        float m_X = 0;
+        float m_Y = 0;
+        float m_Z = 0;
+        float m_Roll = 0;
         SolarObjects m_CenterOfOrbit;
         SolarObjects m_ObjectType;
